Fixed race condition in EfMappingFactory
If 2 threads reached if (!cache.TryGetValue(type, out mapping)) before an item was added to the cache both threads would then try to add the item to the cache causing the second to fail.
By adding a lock and a second TryGetValue we can prevent this by allowing the first thread to successfully add the item to cache then release the lock.
When the second thread is no longer blocked it will do a TryGetValue and retrieve the value from cache.